% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBlast.R
\name{runBlast}
\alias{runBlast}
\title{Perform short sequence alignment with BLAST}
\usage{
runBlast(
  sequences,
  blast_index,
  oppositeStrandOnly = TRUE,
  codingOnly = TRUE,
  word_size = 11,
  style = c("gencode", "ensembl"),
  blastn_bin = NULL
)
}
\arguments{
\item{sequences}{Character vector of DNA sequences.}

\item{blast_index}{String specifying path to a BLAST index.}

\item{oppositeStrandOnly}{If TRUE, only hits for which the target is 
complementary to the query sequences will be returned.}

\item{codingOnly}{If TRUE, only hits for which the target is 
located in a coding gene are returned.}

\item{word_size}{Integer specifying the BLAST word size parameter.
11 by default.}

\item{style}{Either "gencode" (default) or "ensembl".}

\item{blast_bin}{String specifying the path of the BLAST binary program.}
}
\value{
A data.frame of the alignments.
}
\description{
Perform short sequence alignment with BLAST
}
\author{
Jean-Philippe Fortin
}
